Two more fixes for our Wheel CI
1. Add one more explicit dtype annotation to avoid dtype errors on Windows
2. Skip Pallas:SC debug assert tests if not running under Bazel. The tests
   produce errors that put the device in an irrecoverable state without
   restarting the process, which doesn't happen under pytest.

PiperOrigin-RevId: 801879705

+++ jax/_src/numpy/lax_numpy.py
@@ -7633,7 +7633,7 @@ def delete(
     obj = sort(obj)
     obj -= arange(len(obj), dtype=obj.dtype)  # type: ignore
     i = arange(a.shape[axis] - obj.size, dtype=obj.dtype)
-    i += (i[None, :] >= obj[:, None]).sum(0)
+    i += (i[None, :] >= obj[:, None]).sum(0, dtype=i.dtype)
     return a[(slice(None),) * axis + (i,)]
 
   # Case 3b: non-unique indices: must be static.
